+++ src/main/java/com/vietnam/lottery/business/acting/response/ActingListResponse.java
@@ -4,7 +4,6 @@ import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 
 @Data
 public class ActingListResponse implements Serializable {
@@ -17,5 +16,5 @@ public class ActingListResponse implements Serializable {
     private String level;
 
     @ApiModelProperty(value = "分佣比例")
-    private BigDecimal commissionRatio;
+    private Long commissionRatio;
 }
