+++ src/main/java/com/vietnam/lottery/business/grabRedPackets/request/UpdateRequest.java
@@ -5,7 +5,6 @@ import lombok.Data;
 
 import javax.validation.constraints.NotNull;
 import java.io.Serializable;
-import java.math.BigDecimal;
 
 @Data
 public class UpdateRequest implements Serializable {
@@ -16,7 +15,7 @@ public class UpdateRequest implements Serializable {
     private Long id;
 
     @ApiModelProperty(value = "抢红包下注金额")
-    private BigDecimal amount;
+    private Long amount;
 
     @ApiModelProperty(value = "区间开始值")
     private Integer begin;
