+++ src/main/java/com/vietnam/lottery/business/unpackRedPackets/response/UnPackDetailResponse.java
@@ -4,7 +4,6 @@ import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 
 @Data
 public class UnPackDetailResponse implements Serializable {
@@ -20,5 +19,5 @@ public class UnPackDetailResponse implements Serializable {
     private Integer end;
 
     @ApiModelProperty(value = "中奖概率")
-    private BigDecimal probability;
+    private Long probability;
 }
