+++ src/main/java/com/vietnam/lottery/business/unpackRedPackets/response/UnPackListResponse.java
@@ -4,7 +4,6 @@ import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 
 @Data
 public class UnPackListResponse implements Serializable {
@@ -23,5 +22,5 @@ public class UnPackListResponse implements Serializable {
     private Integer end;
 
     @ApiModelProperty(value = "中奖概率")
-    private BigDecimal probability;
+    private Long probability;
 }
